+++ src/gleis/kreuzung.rs
@@ -11,7 +11,7 @@ pub use crate::gleis::weiche::gerade::{
 use crate::{
     gleis::{gerade, kurve, verbindung::Verbindung},
     nachschlagen::impl_nachschlagen,
-    steuerung,
+    steuerung::{self, weiche::MitRichtung},
     typen::{
         canvas::{
             pfad::{self, Pfad, Transformation},
@@ -22,7 +22,7 @@ use crate::{
         skalar::Skalar,
         vektor::Vektor,
         winkel::{self, Trigonometrie, Winkel},
-        MitName, MitRichtung, Transparenz, Zeichnen,
+        MitName, Transparenz, Zeichnen,
     },
 };