+++ src/gleis/weiche/dreiwege.rs
@@ -8,7 +8,10 @@ use zugkontrolle_macros::{alias_serialisiert_unit, erstelle_richtung};
 use crate::{
     gleis::{gerade, kurve, verbindung::Verbindung},
     nachschlagen::impl_nachschlagen,
-    steuerung::{self, weiche::WeicheSteuerung},
+    steuerung::{
+        self,
+        weiche::{MitRichtung, WeicheSteuerung},
+    },
     typen::{
         canvas::{
             pfad::{self, Pfad, Transformation},
@@ -19,7 +22,7 @@ use crate::{
         skalar::Skalar,
         vektor::Vektor,
         winkel::{self, Trigonometrie, Winkel},
-        MitName, MitRichtung, Transparenz, Zeichnen,
+        MitName, Transparenz, Zeichnen,
     },
 };